feat(base): Extract `ObservableMap` into its own module.
